ref: add-feed WIP
