+++ src/main/java/ru/yandex/practicum/filmorate/model/Event.java
@@ -7,11 +7,16 @@ import lombok.*;
 @ToString
 @EqualsAndHashCode
 @NoArgsConstructor
+@RequiredArgsConstructor
 public class Event {
     private Long eventId;
     private Long timestamp;
+    @NonNull
     private EventType eventType;
+    @NonNull
     private OperationType operation;
+    @NonNull
     private Long userId;
+    @NonNull
     private Long entityId;
 }
