+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -5,7 +5,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import ru.yandex.practicum.filmorate.exceptions.ValidationException;
+import ru.yandex.practicum.filmorate.model.Event;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.storages.EventStorage;
 import ru.yandex.practicum.filmorate.storages.FilmStorage;
 import ru.yandex.practicum.filmorate.storages.UserStorage;
 
@@ -26,6 +28,7 @@ import static ru.yandex.practicum.filmorate.model.OperationType.REMOVE;
 public class FilmService {
     private final FilmStorage filmStorage;
     private final UserStorage userStorage;
+    private final EventStorage eventStorage;
     private final DirectorService directorService;
 
     private static void validate(Film film) {
@@ -57,13 +60,13 @@ public class FilmService {
     public void addLike(long filmId, long userId) {
         checkUserExists(userId);
         userStorage.addLike(filmId, userId);
-        userStorage.addEvent(LIKE, ADD, userId, filmId);
+        eventStorage.addEvent(new Event(LIKE, ADD, userId, filmId));
     }
 
     public void removeLike(long filmId, long userId) {
         checkUserExists(userId);
         userStorage.removeLike(filmId, userId);
-        userStorage.addEvent(LIKE, REMOVE, userId, filmId);
+        eventStorage.addEvent(new Event(LIKE, REMOVE, userId, filmId));
     }
 
     public List<Film> getPopular(int count) {
