+++ src/main/java/ru/yandex/practicum/filmorate/service/ReviewService.java
@@ -3,9 +3,10 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import ru.yandex.practicum.filmorate.model.Event;
 import ru.yandex.practicum.filmorate.model.Review;
+import ru.yandex.practicum.filmorate.storages.EventStorage;
 import ru.yandex.practicum.filmorate.storages.ReviewStorage;
-import ru.yandex.practicum.filmorate.storages.UserStorage;
 
 import java.util.List;
 
@@ -17,7 +18,7 @@ import static ru.yandex.practicum.filmorate.model.OperationType.*;
 public class ReviewService {
     private final FilmService filmService;
     private final ReviewStorage reviewStorage;
-    private final UserStorage userStorage;
+    private final EventStorage eventStorage;
 
     public Review getReviewById(int reviewId) {
         return reviewStorage.getReviewById(reviewId);
@@ -35,7 +36,8 @@ public class ReviewService {
         filmService.getFilm(review.getFilmId());
         filmService.checkUserExists(review.getUserId());
         Review returnedReview = reviewStorage.addReview(review);
-        userStorage.addEvent(REVIEW, ADD, review.getUserId(), returnedReview.getReviewId());
+        eventStorage.addEvent(
+                new Event(REVIEW, ADD, review.getUserId().longValue(), (long)returnedReview.getReviewId()));
         return returnedReview;
     }
 
@@ -47,13 +49,14 @@ public class ReviewService {
 
     public Review updateReview(Review review) {
         Review returnedReview = reviewStorage.updateReview(review);
-        userStorage.addEvent(REVIEW, UPDATE, returnedReview.getUserId(), returnedReview.getReviewId());
+        eventStorage.addEvent(
+                new Event(REVIEW, UPDATE, returnedReview.getUserId().longValue(), (long)returnedReview.getReviewId()));
         return returnedReview;
     }
 
     public void removeReview(int reviewId) {
         Review returnedReview = getReviewById(reviewId);
-        userStorage.addEvent(REVIEW, REMOVE, returnedReview.getUserId(), reviewId);
+        eventStorage.addEvent(new Event(REVIEW, REMOVE, returnedReview.getUserId().longValue(), (long)reviewId));
         reviewStorage.removeReview(reviewId);
     }
