+++ src/main/java/ru/yandex/practicum/filmorate/storages/db/UserDbStorage.java
@@ -10,14 +10,10 @@ import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
 
 import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
-import ru.yandex.practicum.filmorate.model.Event;
-import ru.yandex.practicum.filmorate.model.EventType;
-import ru.yandex.practicum.filmorate.model.OperationType;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storages.UserStorage;
 
 import java.sql.Date;
-import java.time.Instant;
 import java.sql.PreparedStatement;
 import java.sql.Statement;
 
@@ -127,35 +123,8 @@ public class UserDbStorage implements UserStorage {
     }
 
     @Override
-    public List<Event> getUserEvents(long userId) {
-        log.debug("/getUserEvents");
-        String sql = "SELECT * FROM events AS e WHERE e.user_id = " + userId + " ORDER BY e.timestamp";
-        return jdbcTemplate.query(sql, rs -> {
-            List<Event> events = new ArrayList<>();
-            while (rs.next()) {
-                Event event = new Event();
-                event.setEventId(rs.getLong("id"));
-                event.setTimestamp(rs.getLong("timestamp"));
-                event.setEventType(EventType.valueOf(rs.getString("event_type")));
-                event.setOperation(OperationType.valueOf(rs.getString("operation")));
-                event.setUserId(rs.getLong("user_id"));
-                event.setEntityId(rs.getLong("entity_id"));
-                events.add(event);
-            }
-            return events;
-        });
-    }
-
-    @Override
     public void deleteById(Long id) {
         String sqlQuery = "DELETE FROM users where id = ?";
         jdbcTemplate.update(sqlQuery, id);
-    }
-
-    @Override
-    public void addEvent(EventType eventType, OperationType operation, long userId, long entityId) {
-        log.debug("/addEvent");
-        String sql = "INSERT INTO events (timestamp, event_type, operation, user_id, entity_id) VALUES (?, ?, ?, ?, ?)";
-        jdbcTemplate.update(sql, Instant.now().toEpochMilli(), eventType.name(), operation.name(), userId, entityId);
     }
 }
