+++ src/main/resources/schema.sql
@@ -86,8 +86,8 @@ CREATE TABLE IF NOT EXISTS review_like_dislike
 create TABLE IF NOT EXISTS events
  (
     id BIGINT generated by default as identity primary key,
-    --timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-    timestamp BIGINT NOT NULL,
+    timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+    --timestamp BIGINT NOT NULL,
     event_type varchar(50) NOT NULL,
     operation varchar(50) NOT NULL,
     user_id BIGINT REFERENCES users(id) ON DELETE CASCADE NOT NULL,
